refactor(docx-reader): move hyperlink and image collection into their own modules
